updated the product model for price to use BigDecimal in place for float

+++ ECOM-BACKEND-1/web/src/main/java/com/ecom/web/model/Product.java
@@ -1,5 +1,7 @@
 package com.ecom.web.model;
 
+import java.math.BigDecimal;
+
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
@@ -7,6 +9,8 @@ import jakarta.persistence.Id;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+ 
+
 
 @Data
 @Entity
@@ -22,6 +26,5 @@ public class Product {
     private String prodName;
     private boolean isAvailable;
     private String prodDescription;
-    private float price;
-    
+    private BigDecimal price;
 }
